% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alabama.R
\name{AL_sales}
\alias{AL_sales}
\title{Alabama Sales Tax}
\usage{
AL_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Alabama Sales Tax
}
