% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arkansas.R
\name{AR_sales}
\alias{AR_sales}
\title{Arkansas Sales Tax}
\usage{
AR_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Arkansas Sales Tax
}
