% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arizona.R
\name{AZ_sales}
\alias{AZ_sales}
\title{Arizona Sales Tax}
\usage{
AZ_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Arizona Sales Tax
}
