% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/California.R
\name{CA_sales}
\alias{CA_sales}
\title{California Sales Tax}
\usage{
CA_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the California Sales Tax
}
