% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colorado.R
\name{CO_sales}
\alias{CO_sales}
\title{Colorado Sales Tax}
\usage{
CO_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Colorado Sales Tax
}
