% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connecticut.R
\name{CT_sales}
\alias{CT_sales}
\title{Connecticut Sales Tax}
\usage{
CT_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Connecticut Sales Tax
}
