% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/District of Columbia.R
\name{DC_sales}
\alias{DC_sales}
\title{District of Columbia Sales Tax}
\usage{
DC_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the District of Columbia Sales Tax
}
