% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Florida.R
\name{FL_sales}
\alias{FL_sales}
\title{Florida Sales Tax}
\usage{
FL_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Florida Sales Tax
}
