% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Georgia.R
\name{GA_sales}
\alias{GA_sales}
\title{Georgia Sales Tax}
\usage{
GA_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Georgia Sales Tax
}
