% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hawaii.R
\name{HI_sales}
\alias{HI_sales}
\title{Hawaii Sales Tax}
\usage{
HI_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Hawaii Sales Tax
}
