% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Idaho.R
\name{ID_sales}
\alias{ID_sales}
\title{Idaho Sales Tax}
\usage{
ID_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Idaho Sales Tax
}
