% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Illinois.R
\name{IL_sales}
\alias{IL_sales}
\title{Illinois Sales Tax}
\usage{
IL_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Illinois Sales Tax
}
