% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indiana.R
\name{IN_sales}
\alias{IN_sales}
\title{Indiana Sales Tax}
\usage{
IN_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Indiana Sales Tax
}
