% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kentucky.R
\name{KY_sales}
\alias{KY_sales}
\title{Kentucky Sales Tax}
\usage{
KY_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Kentucky Sales Tax
}
