% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Louisiana.R
\name{LA_sales}
\alias{LA_sales}
\title{Louisiana Sales Tax}
\usage{
LA_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Louisiana Sales Tax
}
