% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Massachusetts.R
\name{MA_sales}
\alias{MA_sales}
\title{Massachusetts Sales Tax}
\usage{
MA_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Massachusetts Sales Tax
}
