% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maryland.R
\name{MD_sales}
\alias{MD_sales}
\title{Maryland Sales Tax}
\usage{
MD_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Maryland Sales Tax
}
