% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maine.R
\name{ME_sales}
\alias{ME_sales}
\title{Maine Sales Tax}
\usage{
ME_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Maine Sales Tax
}
