% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minnesota.R
\name{MN_sales}
\alias{MN_sales}
\title{Minnesota Sales Tax}
\usage{
MN_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Minnesota Sales Tax
}
