% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Missouri.R
\name{MO_sales}
\alias{MO_sales}
\title{Missouri Sales Tax}
\usage{
MO_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Missouri Sales Tax
}
