% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mississippi.R
\name{MS_sales}
\alias{MS_sales}
\title{Mississippi Sales Tax}
\usage{
MS_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Mississippi Sales Tax
}
