% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/North Carolina.R
\name{NC_sales}
\alias{NC_sales}
\title{North Carolina Sales Tax}
\usage{
NC_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the North Carolina Sales Tax
}
