% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/North Dakota.R
\name{ND_sales}
\alias{ND_sales}
\title{North Dakota Sales Tax}
\usage{
ND_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the North Dakota Sales Tax
}
