% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nebraska.R
\name{NE_sales}
\alias{NE_sales}
\title{Nebraska Sales Tax}
\usage{
NE_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Nebraska Sales Tax
}
