% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New Mexico.R
\name{NM_sales}
\alias{NM_sales}
\title{New Mexico Sales Tax}
\usage{
NM_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the New Mexico Sales Tax
}
