% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nevada.R
\name{NV_sales}
\alias{NV_sales}
\title{Nevada Sales Tax}
\usage{
NV_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Nevada Sales Tax
}
