% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New York.R
\name{NY_sales}
\alias{NY_sales}
\title{New York Sales Tax}
\usage{
NY_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the New York Sales Tax
}
