% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ohio.R
\name{OH_sales}
\alias{OH_sales}
\title{Ohio Sales Tax}
\usage{
OH_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Ohio Sales Tax
}
