% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oklahoma.R
\name{OK_sales}
\alias{OK_sales}
\title{Oklahoma Sales Tax}
\usage{
OK_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Oklahoma Sales Tax
}
