% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pennsylvania.R
\name{PA_sales}
\alias{PA_sales}
\title{Pennsylvania Sales Tax}
\usage{
PA_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Pennsylvania Sales Tax
}
