% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Puerto Rico.R
\name{PR_sales}
\alias{PR_sales}
\title{Puerto Rico Sales Tax}
\usage{
PR_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Puerto Rico Sales Tax
}
