% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhode Island.R
\name{RI_sales}
\alias{RI_sales}
\title{Rhode Island Sales Tax}
\usage{
RI_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Rhode Island Sales Tax
}
