% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/South Carolina.R
\name{SC_sales}
\alias{SC_sales}
\title{South Carolina Sales Tax}
\usage{
SC_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the South Carolina Sales Tax
}
