% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sotuh Dakota.R
\name{SD_sales}
\alias{SD_sales}
\title{South Dakota Sales Tax}
\usage{
SD_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the South Dakota Sales Tax
}
