% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tennessee.R
\name{TN_sales}
\alias{TN_sales}
\title{Tennessee Sales Tax}
\usage{
TN_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Tennessee Sales Tax
}
