% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Texas.R
\name{TX_sales}
\alias{TX_sales}
\title{Texas Sales Tax}
\usage{
TX_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Texas Sales Tax
}
