% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utah.R
\name{UT_sales}
\alias{UT_sales}
\title{Utah Sales Tax}
\usage{
UT_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Utah Sales Tax
}
