% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Virginia.R
\name{VA_sales}
\alias{VA_sales}
\title{Virginia Sales Tax}
\usage{
VA_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Virginia Sales Tax
}
