% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vermont.R
\name{VT_sales}
\alias{VT_sales}
\title{Vermont Sales Tax}
\usage{
VT_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Vermont Sales Tax
}
