% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Washington.R
\name{WA_sales}
\alias{WA_sales}
\title{Washington Sales Tax}
\usage{
WA_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Washington Sales Tax
}
