% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wisconsin.R
\name{WI_sales}
\alias{WI_sales}
\title{Wisconsin Sales Tax}
\usage{
WI_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Wisconsin Sales Tax
}
