% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/West Virginia.R, R/West Virignia.R
\name{WV_sales}
\alias{WV_sales}
\title{West Virginia Sales Tax}
\usage{
WV_sales(x)

WV_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax

total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the West Virginia Sales Tax

Takes any value that is inputted and multiples it by the West Virginia Sales Tax
}
