% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wyoming.R
\name{WY_sales}
\alias{WY_sales}
\title{Wyoming Sales Tax}
\usage{
WY_sales(x)
}
\arguments{
\item{x}{is the value before sales tax}
}
\value{
total with sales tax
}
\description{
Takes any value that is inputted and multiples it by the Wyoming Sales Tax
}
